+++ src/muncher/backgrounds.js
@@ -6,12 +6,13 @@ import { getCampaignId } from "../lib/Settings.js";
 import FileHelper from "../lib/FileHelper.js";
 import SETTINGS from "../settings.js";
 import DDBProxy from "../lib/DDBProxy.js";
+import PatreonHelper from "../lib/PatreonHelper.js";
 
 function getBackgroundData() {
   const cobaltCookie = getCobalt();
   const campaignId = getCampaignId();
   const parsingApi = DDBProxy.getProxy();
-  const betaKey = game.settings.get(SETTINGS.MODULE_ID, "beta-key");
+  const betaKey = PatreonHelper.getPatreonKey();
   const body = { cobalt: cobaltCookie, campaignId: campaignId, betaKey: betaKey };
   const debugJson = game.settings.get(SETTINGS.MODULE_ID, "debug-json");